requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
diff-match-patch==20241021
Django==5.2.7
django-import-export==4.3.13
idna==3.11
requests==2.32.5
sqlparse==0.5.3
tablib==3.9.0
tzdata==2025.2
urllib3==2.5.0
